requirements.txt
pypdf
flask
python-dotenv
sentence-transformers
langchain
langchain-google-genai
langchain-pinecone
pinecone-client>=5.0.0,<6.0.0
-e .